Added some more data classes for the serialiser

+++ src/main/kotlin/uk/tojourn/data/EventBody.kt
@@ -4,6 +4,6 @@ data class EventBody(val eventId: String,
                      val category: String,
                      val subCategory : String,
                      val name: String,
-                     val startTime: Int,
+                     val startTime: Long,
                      val displayed: Boolean,
                      val suspended: Boolean)
